+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Current (in progress)
 
 - Update homepage wording [#617](https://github.com/etalab/udata-gouvfr/pull/617)
+- Change banner trigger for geo.data.gouv.fr [#618](https://github.com/etalab/udata-gouvfr/pull/618)
 
 ## 3.0.5 (2021-08-12)